Migrate to AndroidX. Remove legacy code
