+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,6 @@
             </intent-filter>
         </activity>
         <activity android:name=".SearchActivity"></activity>
-        <activity android:name=".SearchActivityRV"></activity>
     </application>
 
 </manifest>
